"Add category object to CreateProject component and display category name in order summary on create-project view"
This diff adds a new public property `$get_categorie` to the `CreateProject` Livewire component and assigns it the category object corresponding to the selected category in the `updatedCategorieId` method. In the `create-project.blade.php` view, the category name is now displayed in the order summary section if a category is selected.